added new ParametersAlgo.OBS_PROB = MLP_fuzzy and new class MLP_fuzzy
added ParametersAlgo.USE_PERSISTENT_* 
to load the detected posteriograms
PPG and circumvent the slow loading of the MLP model itself with pdnn.

However it does not work for ParametersAlgo.OBS_PROB = MLP_fuzzy,
because we save persistently the log of the posteriograms, but
posteriograms before log are needed to train fuzzy GMMs.